Check max package size for USPS shipping
USPS has a maximum shipping size and this CLI app asks for the packages
dimensions then tells you if the package is too big (or not) to ship
via the US Postal Service. Built for my own needs, but it works for
anyone who wants to ship via USPS. Self contained program, written in
Python3.